api/pyproject.toml
[project]
name = "streamlink-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
"annotated-doc==0.0.4",
"annotated-types==0.7.0",
"anyio==4.12.0",
"attrs==25.4.0",
"certifi==2025.11.12",
"charset-normalizer==3.4.4",
"click==8.3.1",
"fastapi==0.127.0",
"h11==0.16.0",
"idna==3.11",
"isodate==0.7.2",
"lxml==6.0.2",
"outcome==1.3.0.post0",
"pycountry==24.6.1",
"pycryptodome==3.23.0",
"pydantic==2.12.5",
"pydantic_core==2.41.5",
"PySocks==1.7.1",
"requests==2.32.5",
"sniffio==1.3.1",
"sortedcontainers==2.4.0",
"starlette==0.50.0",
"streamlink==8.1.0",
"trio==0.32.0",
"trio-websocket==0.12.2",
"typing-inspection==0.4.2",
"typing_extensions==4.15.0",
"urllib3==2.6.2",
"uvicorn==0.40.0",
"websocket-client==1.9.0",
"python-dotenv==1.0.0",
"supabase==2.11.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
